LPS-31259 Move tests to Layout package and change them to use already existing Util classes
