+++ portal-impl/test/integration/com/liferay/portal/layout/BaseLayoutSetPrototypeTestCase.java
@@ -12,20 +12,14 @@
  * details.
  */
 
-package com.liferay.portal.lar;
+package com.liferay.portal.layout;
 
-import com.liferay.portal.NoSuchLayoutException;
-import com.liferay.portal.kernel.util.FriendlyURLNormalizerUtil;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.Layout;
 import com.liferay.portal.model.LayoutConstants;
-import com.liferay.portal.model.LayoutPrototype;
 import com.liferay.portal.model.LayoutTypePortlet;
 import com.liferay.portal.service.LayoutLocalServiceUtil;
 import com.liferay.portal.service.LayoutServiceUtil;
-import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.service.ServiceTestUtil;
 import com.liferay.portal.util.TestPropsValues;
 
 import org.powermock.api.mockito.PowerMockito;
@@ -33,50 +27,7 @@ import org.powermock.api.mockito.PowerMockito;
 /**
  * @author Eduardo Garcia
  */
-public class BaseExportImportTestCase extends PowerMockito {
-
-	protected Layout addLayout(
-			long groupId, String name, LayoutPrototype layoutPrototype,
-			boolean linkEnabled)
-		throws Exception {
-
-		String friendlyURL =
-			StringPool.SLASH + FriendlyURLNormalizerUtil.normalize(name);
-
-		Layout layout = null;
-
-		try {
-			layout = LayoutLocalServiceUtil.getFriendlyURLLayout(
-				groupId, false, friendlyURL);
-
-			return layout;
-		}
-		catch (NoSuchLayoutException nsle) {
-		}
-
-		ServiceContext serviceContext = ServiceTestUtil.getServiceContext();
-
-		serviceContext.setAttribute("layoutPrototypeLinkEnabled", linkEnabled);
-		serviceContext.setAttribute(
-			"layoutPrototypeUuid", layoutPrototype.getUuid());
-
-		return LayoutLocalServiceUtil.addLayout(
-			TestPropsValues.getUserId(), groupId, false,
-			LayoutConstants.DEFAULT_PARENT_LAYOUT_ID, name, null,
-			"This is a test page.", LayoutConstants.TYPE_PORTLET, false,
-			friendlyURL, serviceContext);
-	}
-
-	protected String addPortlet(
-		Layout layout, String portletId, String columnId) throws Exception {
-		LayoutTypePortlet layoutTypePortlet =
-			(LayoutTypePortlet) layout.getLayoutType();
-
-		portletId = layoutTypePortlet.addPortletId(
-				TestPropsValues.getUserId(), portletId, columnId, 0);
-
-		return portletId;
-	}
+public class BaseLayoutSetPrototypeTestCase extends PowerMockito {
 
 	protected void propagateChanges(Group group) throws Exception {
 		LayoutLocalServiceUtil.getLayouts(
